Début implémentation couche Accès aux données.

+++ src/main/java/fr/djstechno/sensibilisationclimatspringboot/controllers/QuizController.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RestController;
 
-import fr.djstechno.sensibilisationclimatspringboot.models.Quiz;
+import fr.djstechno.sensibilisationclimatspringboot.entities.Quiz;
 import fr.djstechno.sensibilisationclimatspringboot.services.QuizService;
 
 @CrossOrigin(origins = "*")
@@ -24,7 +24,7 @@ public class QuizController {
     }
 
     @GetMapping("/quizs/{id}")
-    public Quiz getQuiz(@PathVariable int id) {
+    public Quiz getQuiz(@PathVariable Long id) {
         return this.quizService.getQuiz(id);
     }
