+++ src/main/java/fr/djstechno/sensibilisationclimatspringboot/services/QuizService.java
@@ -3,12 +3,12 @@ package fr.djstechno.sensibilisationclimatspringboot.services;
 import java.util.List;
 
 import fr.djstechno.sensibilisationclimatspringboot.models.Question;
-import fr.djstechno.sensibilisationclimatspringboot.models.Quiz;
+import fr.djstechno.sensibilisationclimatspringboot.entities.Quiz;
 
 public interface QuizService {
     List<Quiz> getQuizs();
 
-    Quiz getQuiz(int id);
+    Quiz getQuiz(Long id);
 
     List<Question> getQuizQuestions(int id);
 }
